Add files via upload
Task 6: For an array of integers, use slice() to get a new array containing the 2nd, 3rd, and 4th elements.

Task 7: Create an array of names. Use filter() to produce a new array that contains names starting with the letter ‘A’.

Task 8: For an array of scores (out of 100), use map() to grade each score (e.g., 90-100 = ‘A’,80-89 = ‘B’).

Task 9: Given an array of ages, use reduce() to find the average age.

Task 10: Use splice() to insert two new fruits after the second fruit in an array of fruits.